--- src/main/java/db/data/BinIndex.java
@@ -1,10 +1,0 @@
-package db.data;
-
-
-/*
- * 
- */
-public class BinIndex {
-	public Object associatedValue;
-	public int binIndexOnFile;
-}
