--- src/main/java/db/data/BinIndexArrayList.java
@@ -1,12 +1,0 @@
-package db.data;
-
-import java.util.ArrayList;
-
-/**
- * Class used to store 
- *
- */
-@SuppressWarnings("serial")
-public class BinIndexArrayList extends ArrayList<BinIndex> {
-	
-}
