+++ src/main/java/db/data/DoubleType.java
@@ -4,6 +4,8 @@ import java.nio.ByteBuffer;
 
 import org.apache.commons.lang3.ArrayUtils;
 
+import db.search.Operator;
+
 
 public class DoubleType extends DataType {
