+++ src/main/java/db/data/StringType.java
@@ -4,9 +4,8 @@ import java.nio.ByteBuffer;
 import java.util.Arrays;
 
 import org.apache.commons.lang3.ArrayUtils;
-import org.apache.commons.lang3.StringUtils;
 
-import com.dant.utils.Log;
+import db.search.Operator;
 
 public class StringType extends DataType {
 	public static boolean sizeIsRequired = true;
