+++ src/main/java/db/search/Field.java
@@ -1,4 +1,4 @@
-package db.data;
+package db.search;
 
 public class Field {
 	protected String name;
