+++ src/main/java/db/search/Operable.java
@@ -1,4 +1,4 @@
-package db.data;
+package db.search;
 
 public interface Operable {	
 	public boolean isOperatorCompatible(Operator op);
