+++ src/main/java/db/search/Operator.java
@@ -1,4 +1,4 @@
-package db.data;
+package db.search;
 
 public enum Operator {
 	equals,
