+++ src/main/java/db/structure/Database.java
@@ -7,11 +7,9 @@ import com.google.gson.Gson;
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import javax.ws.rs.core.Response;
-import com.dant.utils.Utils;
 
 public final class Database {
